Refactor Intervals router
- Separate OAuth and webhook handlers into distinct modules
- Implement decorator for independent athlete processing
- Streamline OAuth callback for token exchange
- Modify scheduler to decouple wellness sync
- Adjust webhook to support wellness event dispatch
