+++ tasks/actors/reports.py
@@ -182,12 +182,7 @@ def actor_compose_user_morning_report(
     _dt = DateDTO.today().isoformat()
 
     with get_sync_session() as session:
-        _wellness_row = session.execute(
-            select(Wellness).where(
-                Wellness.user_id == user.id,
-                Wellness.date == _dt,
-            )
-        ).scalar_one_or_none()
+        _wellness_row = Wellness.get(user_id=user.id, dt=_dt, session=session)
 
         if not _wellness_row or not _wellness_row.sleep_score or _wellness_row.ai_recommendation:
             return
